requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
h11==0.12.0
idna==2.10
numpy==1.20.2
Pillow==8.2.0
pydantic==1.8.1
python-multipart==0.0.5
requests==2.25.1
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
